refactor: AppLoader and init folder
