+++ src/components/App/index.js
@@ -1,0 +1,2 @@
+// @flow
+export { default as App } from './App.component';
