+++ src/components/AppLoader/index.js
@@ -1,0 +1,2 @@
+// @flow
+export { default as AppLoader } from './AppLoader.component';
