+++ src/entry/addRxjsOperators.js
