+++ src/entry/index.js
@@ -1,0 +1,4 @@
+// @flow
+export { triggerLoadCoreEpic, loadAppEpic, loadCoreFailedEpic } from './entry.epics';
+export { actionTypes } from './entry.actions';
+export { initApp } from './entry';
