+++ src/locales/en/translations.json
@@ -1,5 +1,8 @@
 {
-    "Please provide a valid date": "",
+    "Choose one or more dates...": "",
+    "Choose a date...": "",
+    "Today": "",
+    "A possible reason for this is that the browser or mode (e.g. privacy mode) is not supported. See log for details.": "",
     "Columns to show in table": "",
     "Column": "",
     "This value is validating": "",
@@ -10,6 +13,7 @@
     "Please provide a positive integer": "",
     "Please provide zero or a positive integer": "",
     "Please provide a negative integer": "",
+    "Please provide a valid date": "",
     "Please provide a valid date and time": "",
     "Please provide a valid time": "",
     "Please provide a valid percentage": "",
@@ -61,7 +65,6 @@
     "to": "",
     "From": "",
     "To": "",
-    "Today": "",
     "This week": "",
     "This month": "",
     "This Year": "",
@@ -251,8 +254,5 @@
     "Delete polygon": "",
     "Set area": "",
     "You opened another version of the Capture App in the same domain. Currently, the Capture App only supports running one version concurrently (in the same domain). Please refresh this page if you would like to use this version again, but be aware that this will close other versions.": "",
-    "Choose one or more dates...": "",
-    "Choose a date...": "",
-    "A possible reason for this is that the browser or mode (e.g. privacy mode) is not supported. See log for details.": "",
     "Compatibility mode": ""
 }
