+++ src/reducers/descriptions/app.reducerDescription.js
@@ -1,6 +1,6 @@
 // @flow
 import { getAppReducerDesc } from 'capture-core/reducers/descriptions/app.reducerDescriptionGetter';
-import { actionTypes as entryActionTypes } from '../../init/entry.actions';
+import { actionTypes as entryActionTypes } from '../../entry';
 
 export const appReducerDesc = getAppReducerDesc({
     [entryActionTypes.APP_LOAD]: state => ({
